bank account1

+++ programming_paradigm/main-0.py
@@ -9,7 +9,7 @@ def main():
         sys.exit(1)
 
     command, *params = sys.argv[1].split(':')
-    amount = int(params[0]) if params else None
+    amount = float(params[0]) if params else None
 
     if command == "deposit" and amount is not None:
         account.deposit(amount)
